qdrant_client

Module qdrant

API types

Modules

alias_operations

Nested message and enum types in AliasOperations.

binary_quantization_query_encoding

Nested message and enum types in BinaryQuantizationQueryEncoding.

collections_client

Generated client implementations.

collections_server

Generated server implementations.

condition

Nested message and enum types in Condition.

count_points_builder

create_collection_builder

dense_vector_builder

document_builder

expression

Nested message and enum types in Expression.

facet_value

Nested message and enum types in FacetValue.

group_id

Nested message and enum types in GroupId.

image_builder

match

Nested message and enum types in Match.

max_optimization_threads

Nested message and enum types in MaxOptimizationThreads.

multi_dense_vector_builder

order_value

Nested message and enum types in OrderValue.

payload_index_params

Nested message and enum types in PayloadIndexParams.

point_id

Nested message and enum types in PointId.

points_client

Generated client implementations.

points_selector

Nested message and enum types in PointsSelector.

points_server

Generated server implementations.

points_update_operation

Nested message and enum types in PointsUpdateOperation.

qdrant_client

Generated client implementations.

qdrant_server

Generated server implementations.

quantization_config

Nested message and enum types in QuantizationConfig.

quantization_config_diff

Nested message and enum types in QuantizationConfigDiff.

query

Nested message and enum types in Query.

read_consistency

Nested message and enum types in ReadConsistency.

shard_key

Nested message and enum types in ShardKey.

snapshots_client

Generated client implementations.

snapshots_server

Generated server implementations.

sparse_vector_builder

sparse_vectors_config

start_from

Nested message and enum types in StartFrom.

stemming_algorithm

Nested message and enum types in StemmingAlgorithm.

target_vector

Nested message and enum types in TargetVector.

update_collection_cluster_setup_request

Nested message and enum types in UpdateCollectionClusterSetupRequest.

value

Nested message and enum types in Value.

vector

Nested message and enum types in Vector.

vector_example

Nested message and enum types in VectorExample.

vector_input

Nested message and enum types in VectorInput.

vector_output

Nested message and enum types in VectorOutput.

vectors

Nested message and enum types in Vectors.

vectors_config

Nested message and enum types in VectorsConfig.

vectors_config_diff

Nested message and enum types in VectorsConfigDiff.

vectors_output

Nested message and enum types in VectorsOutput.

with_payload_selector

Nested message and enum types in WithPayloadSelector.

with_vectors_selector

Nested message and enum types in WithVectorsSelector.

Structs

AbortShardTransfer

AbortShardTransferBuilder

AliasDescription

AliasOperations

BatchResult

BinaryQuantization

BinaryQuantizationBuilder

BinaryQuantizationQueryEncoding

BoolIndexParams

BoolIndexParamsBuilder

ChangeAliases

ClearPayloadPoints

ClearPayloadPointsBuilder

CollectionClusterInfoRequest

CollectionClusterInfoResponse

CollectionConfig

CollectionDescription

CollectionExists

CollectionExistsRequest

CollectionExistsResponse

CollectionInfo

CollectionOperationResponse

CollectionParams

CollectionParamsDiff

CollectionParamsDiffBuilder

Condition

ContextExamplePair

ContextExamplePairBuilder

ContextInput

ContextInputBuilder

ContextInputPair

ContextInputPairBuilder

CountPoints

CountPointsBuilder

CountResponse

CountResult

CreateAlias

CreateAliasBuilder

CreateCollection

CreateCollectionBuilder

CreateFieldIndexCollection

CreateFieldIndexCollectionBuilder

CreateFullSnapshotRequest

CreateShardKey

CreateShardKeyBuilder

CreateShardKeyRequest

CreateShardKeyRequestBuilder

CreateShardKeyResponse

CreateSnapshotRequest

CreateSnapshotResponse

DatetimeIndexParams

DatetimeIndexParamsBuilder

DatetimeRange

DecayParamsExpression

DecayParamsExpressionBuilder

Builder for the DecayParamsExpression struct, which represents parameters for decay functions.

DeleteAlias

DeleteCollection

DeleteCollectionBuilder

DeleteFieldIndexCollection

DeleteFieldIndexCollectionBuilder

DeleteFullSnapshotRequest

DeletePayloadPoints

DeletePayloadPointsBuilder

DeletePointVectors

DeletePointVectorsBuilder

DeletePoints

DeletePointsBuilder

DeleteShardKey

DeleteShardKeyRequest

DeleteShardKeyRequestBuilder

DeleteShardKeyResponse

DeleteSnapshotRequest

DeleteSnapshotRequestBuilder

DeleteSnapshotResponse

DenseVector

DenseVectorBuilder

Disabled

DiscoverBatchPoints

DiscoverBatchPointsBuilder

DiscoverBatchResponse

DiscoverInput

DiscoverInputBuilder

DiscoverPoints

DiscoverPointsBuilder

DiscoverResponse

DivExpression

Document

DocumentBuilder

Expression

FacetCounts

FacetCountsBuilder

FacetHit

FacetResponse

FacetValue

FieldCondition

Filter

FloatIndexParams

FloatIndexParamsBuilder

Formula

FormulaBuilder

Builder for the Formula struct, which represents a scoring formula for points.

GeoBoundingBox

GeoDistance

GeoIndexParams

GeoIndexParamsBuilder

GeoLineString

GeoPoint

GeoPolygon

For a valid GeoPolygon, both the exterior and interior GeoLineStrings must consist of a minimum of 4 points. Additionally, the first and last points of each GeoLineString must be the same.

GeoRadius

GetCollectionInfoRequest

GetCollectionInfoResponse

GetPoints

GetPointsBuilder

GetResponse

GroupId

GroupsResult

HardwareUsage

HasIdCondition

HasVectorCondition

HealthCheckReply

HealthCheckRequest

HnswConfigDiff

HnswConfigDiffBuilder

Image

ImageBuilder

InferenceObject

InferenceUsage

IntegerIndexParams

IntegerIndexParamsBuilder

IsEmptyCondition

IsNullCondition

KeywordIndexParams

KeywordIndexParamsBuilder

ListAliasesRequest

ListAliasesResponse

ListCollectionAliasesRequest

ListCollectionsRequest

ListCollectionsResponse

ListFullSnapshotsRequest

ListSnapshotsRequest

ListSnapshotsResponse

ListValue

ListValue is a wrapper around a repeated field of values.

LocalShardInfo

LookupLocation

LookupLocationBuilder

Match

MaxOptimizationThreads

MaxOptimizationThreadsBuilder

Max number of threads (jobs) for running optimizations per shard. Each optimization job will also use max_indexing_threads threads by itself for index building.

MinShould

Mmr

Maximal Marginal Relevance (MMR) algorithm for re-ranking the points.

MmrBuilder

ModelUsage

MoveShard

MoveShardBuilder

MultExpression

MultiDenseVector

MultiDenseVectorBuilder

MultiVectorConfig

MultiVectorConfigBuilder

NamedVectors

NamedVectorsOutput

NearestInputWithMmr

NestedCondition

OptimizerStatus

OptimizersConfigDiff

OptimizersConfigDiffBuilder

OrderBy

OrderByBuilder

OrderValue

PayloadExcludeSelector

PayloadIncludeSelector

PayloadIndexParams

PayloadSchemaInfo

PointGroup

PointId

PointStruct

PointVectors

PointsIdsList

PointsOperationResponse

PointsSelector

PointsUpdateOperation

PowExpression

PrefetchQuery

PrefetchQueryBuilder

ProductQuantization

ProductQuantizationBuilder

QuantizationConfig

QuantizationConfigDiff

QuantizationSearchParams

QuantizationSearchParamsBuilder

Query

QueryBatchPoints

QueryBatchPointsBuilder

QueryBatchResponse

QueryGroupsResponse

QueryPointGroups

QueryPointGroupsBuilder

QueryPoints

QueryPointsBuilder

QueryResponse

Range

ReadConsistency

RecommendBatchPoints

RecommendBatchPointsBuilder

RecommendBatchResponse

RecommendGroupsResponse

RecommendInput

RecommendInputBuilder

RecommendPointGroups

RecommendPointGroupsBuilder

RecommendPoints

RecommendPointsBuilder

RecommendResponse

RemoteShardInfo

RenameAlias

RenameAliasBuilder

RepeatedIntegers

RepeatedStrings

Replica

ReplicaBuilder

ReplicateShard

ReplicateShardBuilder

ReshardingInfo

RestartTransfer

RetrievedPoint

ScalarQuantization

ScalarQuantizationBuilder

ScoredPoint

ScrollPoints

ScrollPointsBuilder

ScrollResponse

SearchBatchPoints

SearchBatchPointsBuilder

SearchBatchResponse

SearchGroupsResponse

SearchMatrixOffsets

SearchMatrixOffsetsResponse

SearchMatrixPair

SearchMatrixPairs

SearchMatrixPairsResponse

SearchMatrixPoints

SearchMatrixPointsBuilder

SearchParams

SearchParamsBuilder

SearchPointGroups

SearchPointGroupsBuilder

SearchPoints

SearchPointsBuilder

SearchResponse

SetPayloadPoints

SetPayloadPointsBuilder

ShardKey

ShardKeySelector

ShardTransferInfo

SnapshotDescription

SnapshotDownload

SnapshotDownloadBuilder

Builder for SnapshotDownload.

SnowballParams

SparseIndexConfig

SparseIndexConfigBuilder

SparseIndices

SparseVector

SparseVectorBuilder

SparseVectorConfig

SparseVectorParams

SparseVectorParamsBuilder

SparseVectorsConfigBuilder

StartFrom

StemmingAlgorithm

StopwordsSet

StrictModeConfig

StrictModeConfigBuilder

StrictModeMultivector

StrictModeMultivectorConfig

StrictModeMultivectorConfigBuilder

Builder for StrictModeMultivectorConfig, which defines multivector configuration for strict mode.

StrictModeSparse

StrictModeSparseConfig

StrictModeSparseConfigBuilder

Builder for StrictModeSparseConfig, which defines sparse vector configuration for strict mode.

Struct

Struct represents a structured data value, consisting of fields which map to dynamically typed values. In some languages, Struct might be supported by a native representation. For example, in scripting languages like JS a struct is represented as an object. The details of that representation are described together with the proto support for the language.

SumExpression

TargetVector

TextIndexParams

TextIndexParamsBuilder

Timestamp

A Timestamp represents a point in time independent of any time zone or local calendar, encoded as a count of seconds and fractions of seconds at nanosecond resolution. The count is relative to an epoch at UTC midnight on January 1, 1970, in the proleptic Gregorian calendar which extends the Gregorian calendar backwards to year one.

UpdateBatchPoints

UpdateBatchPointsBuilder

UpdateBatchResponse

UpdateCollection

UpdateCollectionBuilder

UpdateCollectionClusterSetupRequest

UpdateCollectionClusterSetupRequestBuilder

UpdateCollectionClusterSetupResponse

UpdatePointVectors

UpdatePointVectorsBuilder

UpdateResult

UpsertPoints

UpsertPointsBuilder

Usage

UuidIndexParams

UuidIndexParamsBuilder

Value

Value represents a dynamically typed value which can be either null, a number, a string, a boolean, a recursive struct value, or a list of values. A producer of value is expected to set one of those variants, absence of any variant indicates an error.

ValuesCount

Vector

Legacy vector format, which determines the vector type by the configuration of its fields.

VectorExample

VectorInput

Vector type to be used in queries. Ids will be substituted with their corresponding vectors from the collection.

VectorOutput

VectorParams

VectorParamsBuilder

VectorParamsDiff

VectorParamsDiffBuilder

VectorParamsDiffMap

VectorParamsMap

Vectors

VectorsConfig

VectorsConfigBuilder

VectorsConfigDiff

VectorsOutput

VectorsSelector

WalConfigDiff

WalConfigDiffBuilder

WithLookup

WithLookupBuilder

WithPayloadSelector

WithVectorsSelector

WriteOrdering

Enums

BinaryQuantizationEncoding

CollectionStatus

CompressionRatio

CountPointsBuilderError

Error type for CountPointsBuilder

Datatype

Direction

Distance

FieldType

Fusion

Modifier

MultiVectorComparator

NullValue

NullValue is a singleton enumeration to represent the null value for the Value type union.

PayloadSchemaType

QuantizationType

ReadConsistencyType

RecommendExample

RecommendStrategy

How to use positive and negative vectors to find the results, default is AverageVector.

ReplicaState

ReshardingDirection

Resharding direction, scale up or down in number of shards

Sample

Sample points from the collection

ShardTransferMethod

ShardingMethod

SnapshotDownloadBuilderError

Error type for SnapshotDownloadBuilder

TokenizerType

UpdateStatus

WriteOrderingType