qdrant_client::qdrant::collections_server

Trait Collections

pub trait Collections:
    Send
    + Sync
    + 'static {

    // Required methods
    fn get<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetCollectionInfoRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetCollectionInfoResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn list<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ListCollectionsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ListCollectionsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn create<'life0, 'async_trait>(
        &'life0 self,
        request: Request<CreateCollection>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CollectionOperationResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn update<'life0, 'async_trait>(
        &'life0 self,
        request: Request<UpdateCollection>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CollectionOperationResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn delete<'life0, 'async_trait>(
        &'life0 self,
        request: Request<DeleteCollection>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CollectionOperationResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn update_aliases<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ChangeAliases>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CollectionOperationResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn list_collection_aliases<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ListCollectionAliasesRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ListAliasesResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn list_aliases<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ListAliasesRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ListAliasesResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn collection_cluster_info<'life0, 'async_trait>(
        &'life0 self,
        request: Request<CollectionClusterInfoRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CollectionClusterInfoResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn collection_exists<'life0, 'async_trait>(
        &'life0 self,
        request: Request<CollectionExistsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CollectionExistsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn update_collection_cluster_setup<'life0, 'async_trait>(
        &'life0 self,
        request: Request<UpdateCollectionClusterSetupRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<UpdateCollectionClusterSetupResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn create_shard_key<'life0, 'async_trait>(
        &'life0 self,
        request: Request<CreateShardKeyRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CreateShardKeyResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn delete_shard_key<'life0, 'async_trait>(
        &'life0 self,
        request: Request<DeleteShardKeyRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<DeleteShardKeyResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

}

Generated trait containing gRPC methods that should be implemented for use with CollectionsServer.

Required Methods

fn get<'life0, 'async_trait>( &'life0 self, request: Request<GetCollectionInfoRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetCollectionInfoResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get detailed information about specified existing collection

fn list<'life0, 'async_trait>( &'life0 self, request: Request<ListCollectionsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ListCollectionsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get list name of all existing collections

fn create<'life0, 'async_trait>( &'life0 self, request: Request<CreateCollection>, ) -> Pin<Box<dyn Future<Output = Result<Response<CollectionOperationResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Create new collection with given parameters

fn update<'life0, 'async_trait>( &'life0 self, request: Request<UpdateCollection>, ) -> Pin<Box<dyn Future<Output = Result<Response<CollectionOperationResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Update parameters of the existing collection

fn delete<'life0, 'async_trait>( &'life0 self, request: Request<DeleteCollection>, ) -> Pin<Box<dyn Future<Output = Result<Response<CollectionOperationResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Drop collection and all associated data

fn update_aliases<'life0, 'async_trait>( &'life0 self, request: Request<ChangeAliases>, ) -> Pin<Box<dyn Future<Output = Result<Response<CollectionOperationResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Update Aliases of the existing collection

fn list_collection_aliases<'life0, 'async_trait>( &'life0 self, request: Request<ListCollectionAliasesRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ListAliasesResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get list of all aliases for a collection

fn list_aliases<'life0, 'async_trait>( &'life0 self, request: Request<ListAliasesRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ListAliasesResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get list of all aliases for all existing collections

fn collection_cluster_info<'life0, 'async_trait>( &'life0 self, request: Request<CollectionClusterInfoRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<CollectionClusterInfoResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get cluster information for a collection

fn collection_exists<'life0, 'async_trait>( &'life0 self, request: Request<CollectionExistsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<CollectionExistsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Check the existence of a collection

fn update_collection_cluster_setup<'life0, 'async_trait>( &'life0 self, request: Request<UpdateCollectionClusterSetupRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<UpdateCollectionClusterSetupResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Update cluster setup for a collection

fn create_shard_key<'life0, 'async_trait>( &'life0 self, request: Request<CreateShardKeyRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<CreateShardKeyResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Create shard key

fn delete_shard_key<'life0, 'async_trait>( &'life0 self, request: Request<DeleteShardKeyRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<DeleteShardKeyResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Delete shard key