qdrant_client::qdrant::payload_index_params

Enum IndexParams

pub enum IndexParams {
    KeywordIndexParams(KeywordIndexParams),
    IntegerIndexParams(IntegerIndexParams),
    FloatIndexParams(FloatIndexParams),
    GeoIndexParams(GeoIndexParams),
    TextIndexParams(TextIndexParams),
    BoolIndexParams(BoolIndexParams),
    DatetimeIndexParams(DatetimeIndexParams),
    UuidIndexParams(UuidIndexParams),
}

Variants

KeywordIndexParams(KeywordIndexParams)

Parameters for keyword index

IntegerIndexParams(IntegerIndexParams)

Parameters for integer index

FloatIndexParams(FloatIndexParams)

Parameters for float index

GeoIndexParams(GeoIndexParams)

Parameters for geo index

TextIndexParams(TextIndexParams)

Parameters for text index

BoolIndexParams(BoolIndexParams)

Parameters for bool index

DatetimeIndexParams(DatetimeIndexParams)

Parameters for datetime index

UuidIndexParams(UuidIndexParams)

Parameters for uuid index

Implementations

impl IndexParams

pub fn encode(&self, buf: &mut impl BufMut)

Encodes the message to a buffer.

pub fn merge( field: &mut Option<IndexParams>, tag: u32, wire_type: WireType, buf: &mut impl Buf, ctx: DecodeContext, ) -> Result<(), DecodeError>

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.