qdrant_client::qdrant

Struct Expression

pub struct Expression {
    pub variant: Option<Variant>,
}

Fields
variant: Option<Variant>
Implementations

impl Expression

pub fn constant(value: f32) -> Self

Creates a new Expression with a constant value.

pub fn variable<S: Into<String>>(name: S) -> Self

Creates a new Expression with a variable (payload key or reference to score).

pub fn score() -> Self

Creates a new Expression with a reference to the score of the prefetch.

pub fn score_idx(idx: usize) -> Self

Creates a new Expression with a reference to the score of a specific prefetch, when there are multiple prefetches.

pub fn condition<C: Into<Condition>>(condition: C) -> Self

Creates a new Expression with a condition. If true, becomes 1.0; otherwise 0.0.

pub fn geo_distance<G: Into<GeoDistance>>(geo_distance: G) -> Self

Creates a new Expression with a geographic distance in meters.

pub fn datetime<S: Into<String>>(datetime: S) -> Self

Creates a new Expression with a date-time constant.

pub fn datetime_key<S: Into<String>>(key: S) -> Self

Creates a new Expression with a payload key with date-time values.

pub fn mult<M: Into<MultExpression>>(mult: M) -> Self

Creates a new Expression with a multiplication expression.

pub fn sum<S: Into<SumExpression>>(sum: S) -> Self

Creates a new Expression with a sum expression.

pub fn div<D: Into<DivExpression>>(div: D) -> Self

Creates a new Expression with a division expression.

pub fn neg<E: Into<Expression>>(expr: E) -> Self

Creates a new Expression with a negation expression.

pub fn abs<E: Into<Expression>>(expr: E) -> Self

Creates a new Expression with an absolute value expression.

pub fn sqrt<E: Into<Expression>>(expr: E) -> Self

Creates a new Expression with a square root expression.

pub fn pow<P: Into<PowExpression>>(pow: P) -> Self

Creates a new Expression with a power expression.

pub fn exp<E: Into<Expression>>(expr: E) -> Self

Creates a new Expression with an exponential expression.

pub fn log10<E: Into<Expression>>(expr: E) -> Self

Creates a new Expression with a log10 expression.

pub fn ln<E: Into<Expression>>(expr: E) -> Self

Creates a new Expression with a natural logarithm expression.

pub fn geo_distance_with<G: Into<GeoPoint>, S: Into<String>>( origin: G, to: S, ) -> Self

Creates a new Expression with a geo distance expression.

pub fn exp_decay<D: Into<DecayParamsExpression>>(decay: D) -> Self

Creates a new Expression with an exponential decay expression.

pub fn gauss_decay<D: Into<DecayParamsExpression>>(decay: D) -> Self

Creates a new Expression with a Gaussian decay expression.

pub fn lin_decay<D: Into<DecayParamsExpression>>(decay: D) -> Self

Creates a new Expression with a linear decay expression.

pub fn mult_with<E: Into<Expression>, I: IntoIterator<Item = E>>( expressions: I, ) -> Self

Helper method to create a multiplication expression with multiple sub-expressions.

pub fn sum_with<E: Into<Expression>, I: IntoIterator<Item = E>>( expressions: I, ) -> Self

Helper method to create a sum expression with multiple sub-expressions.

pub fn div_with<L: Into<Expression>, R: Into<Expression>>( left: L, right: R, by_zero_default: Option<f32>, ) -> Self

Helper method to create a division expression with left and right operands.

pub fn pow_with<B: Into<Expression>, E: Into<Expression>>( base: B, exponent: E, ) -> Self

Helper method to create a power expression with base and exponent.