qdrant_client::qdrant

Struct Filter

pub struct Filter {
    pub should: Vec<Condition>,
    pub must: Vec<Condition>,
    pub must_not: Vec<Condition>,
    pub min_should: Option<MinShould>,
}

Fields
should: Vec<Condition>

At least one of those conditions should match

must: Vec<Condition>

All conditions must match

must_not: Vec<Condition>

All conditions must NOT match

min_should: Option<MinShould>

At least minimum amount of given conditions should match

Implementations

impl Filter

pub fn must(conds: impl IntoIterator<Item = Condition>) -> Self

Create a Filter where all the conditions must be satisfied.

pub fn should(conds: impl IntoIterator<Item = Condition>) -> Self

Create a Filter where at least one of the conditions should be satisfied.

pub fn min_should( min_count: u64, conds: impl IntoIterator<Item = Condition>, ) -> Self

Create a Filter where at least a minimum amount of given conditions should be statisfied.

pub fn must_not(conds: impl IntoIterator<Item = Condition>) -> Self

Create a Filter where none of the conditions must be satisfied.

pub fn any(conds: impl IntoIterator<Item = Condition>) -> Self

Alias for should.

Create a Filter that matches if any of the conditions match.

pub fn all(conds: impl IntoIterator<Item = Condition>) -> Self

Alias for must.

Create a Filter that matches if all of the conditions match.

pub fn none(conds: impl IntoIterator<Item = Condition>) -> Self

Alias for must_not.

Create a Filter that matches if none of the conditions match.